Arrumando imagem

+++ marketplace/src/main/java/com/iroque/marketplace/services/ProductService.java
@@ -1,8 +1,12 @@
 package com.iroque.marketplace.services;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.springframework.stereotype.Service;
+import org.springframework.util.Base64Utils;
 
 import com.iroque.marketplace.interfaces.ProductRepository;
 import com.iroque.marketplace.models.Product;
@@ -22,7 +26,7 @@ public class ProductService {
 
     public Product getProduct(Long id) {
         return productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Product not found with ID: " + id));
+                .orElseThrow(() -> new RuntimeException("Produto não encontrado com o ID: " + id));
     }
 
     public List<Product> getProductByName(String name) {
@@ -34,16 +38,21 @@ public class ProductService {
     }
 
     public Product updateProduct(Long id, Product updateProduct) {
+        byte[] imageBytes = Base64Utils.decodeFromString(new String(updateProduct.getImage()));
         Product product = getProduct(id);
         product.setName(updateProduct.getName());
         product.setDescription(updateProduct.getDescription());
         product.setPrice(updateProduct.getPrice());
         product.setQuantity(updateProduct.getQuantity());
-        product.setImage(updateProduct.getImage());
+        product.setImage(imageBytes);
         return productRepository.save(product);
     }
 
     public void deleteProduct(Long id) {
+        Optional<Product> product = productRepository.findById(id);
+        if (!product.isPresent()) {
+            throw new RuntimeException("Produto com ID " + id + " não encontrado.");
+        }
         productRepository.deleteById(id);
     }
